Simplify task input by moving if/else up
Since the VCF index file cannot be passed explicitly, it has to be in
the same folder as the VCF file itself. This means that it is not
possible to generate the index file in a different task, since then they
will not be localised together.

Therefore, if an index file was specified, we can use the specified VCF
file. However, if an index file was not specified, we need to use both
the VCF file and the index from the indexing task

Instead of putting this logic inside the task inputs, it was moved to
the beginning of the pipeline, so the task input is cleaner.